requirements.txt
elasticsearch
sentence-transformers
faiss-cpu
pandas
numpy
scikit-learn
tqdm
sklearn